Refactor language handling by introducing LangType enum and updating related methods; bump version to 0.3.2.1-STABLE
