+++ readme.md
@@ -17,7 +17,7 @@ And if you thought APIs were complicated, think again! With ConnectLib, it's lik
 ---
 
 ```java
-Stable Version: 0.3.2-STABLE
+Stable Version: 0.3.2.1-STABLE
 ```
 
 ---
@@ -115,7 +115,7 @@ repositories {
 
 dependencies {
     
-    implementation("fr.sandro642.github:ConnectLib:0.3.2-STABLE")
+    implementation("fr.sandro642.github:ConnectLib:0.3.2.1-STABLE")
     
 }
